rec command の処理を recManager へ分割
